Backend/requirements.txt
click
dash-auth
dash-bootstrap-components>=1
dash>=2.3.1
dtreeviz>=1.3
flask_simplelogin
graphviz>=0.18.2
joblib
jupyter_dash
numpy
oyaml
pandas>=1.1
scikit-learn>=1.1
shap>=0.37
waitress
lightgbm
xgboost
django
djangorestframework
django-cors-headers
django-extensions
pandas
